+++ FireAxe.GUI/AppSettingsExtensions.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace FireAxe;
+
+public static class AppSettingsExtensions
+{
+    public static DownloadServiceSettings GetDownloadServiceSettings(this AppSettings appSettings)
+    {
+        ArgumentNullException.ThrowIfNull(appSettings);
+
+        return new DownloadServiceSettings
+        {
+            Proxy = appSettings.WebProxy
+        };
+    }
+}