+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add: addon deletion progress window
 - add: mechanism of addon linked files: all files with the same name as the addon file will follow the addon when it is moved
 - add: alternative method of customizing addon preview image: place an image file with the same name next to the addon file (issue #17)
+- add: web proxy address setting (issue #19)
 - fix: deletion of addons doesn't delete the corresponding files in some cases
 - improve: better download logic of WorkshopVpkAddon
 - improve: move the addon priority setting to the AddonNode view and support addon global priority (now AddonGroups have priority too)